--- app/Domains/Permissions/Feature/Model/Collection/Feature.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace App\Domains\Permissions\Feature\Model\Collection;
-
-use App\Domains\CoreApp\Model\Collection\CollectionAbstract;
-
-class Feature extends CollectionAbstract
-{
-}
