--- app/Domains/Permissions/Model/Collection/Permission.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace App\Domains\Permissions\Model\Collection;
-
-use App\Domains\CoreApp\Model\Collection\CollectionAbstract;
-
-class Permission extends CollectionAbstract
-{
-}
