--- app/Domains/Permissions/Service/Controller/ControllerAbstract.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace App\Domains\Permissions\Service\Controller;
-
-use App\Domains\CoreApp\Service\Controller\ControllerAbstract as ControllerAbstractCore;
-
-abstract class ControllerAbstract extends ControllerAbstractCore
-{
-}
