+++ app/Domains/User/Permissions/Action/ActionFactory.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace App\Domains\User\Permissions\Action;
+
+use App\Domains\User\Permissions\Model\Permission as Model;
+use App\Domains\Core\Action\ActionFactoryAbstract;
+
+class ActionFactory extends ActionFactoryAbstract
+{
+    protected ?Model $row;
+
+    public function create(array $data): Model
+    {
+        return $this->actionHandle(Create::class, $data, $data); // Truyền $data vào $args
+    }
+}
