+++ app/Domains/User/Permissions/Feature/Model/Collection/Feature.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace App\Domains\User\Permissions\Feature\Model\Collection;
+
+use App\Domains\CoreApp\Model\Collection\CollectionAbstract;
+
+class Feature extends CollectionAbstract
+{
+}
