+++ app/Domains/User/Permissions/Model/Collection/Permission.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace App\Domains\User\Permissions\Model\Collection;
+
+use App\Domains\CoreApp\Model\Collection\CollectionAbstract;
+
+class Permission extends CollectionAbstract
+{
+}
