+++ app/Domains/User/Permissions/Model/RolePermission.php
@@ -1,0 +1,22 @@
+<?php declare(strict_types=1);
+
+namespace App\Domains\User\Permissions\Model;
+
+use Illuminate\Database\Eloquent\Model;
+
+class RolePermission extends Model
+{
+    protected $table = 'role_permissions';
+
+    protected $fillable = [
+        'role_id',
+        'permission_id',
+    ];
+
+    public $timestamps = false;
+
+    public function permission()
+    {
+        return $this->belongsTo(Permission::class, 'permission_id');
+    }
+}
